src/ssh-exec/pyproject.toml
[project]
name = "mcp-ssh-exec"
version = "0.1.0"
description = "MCP server for executing command-line operations on remote Linux systems via SSH"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.5.0",
    "paramiko>=3.4.0",
    "pydantic>=2.5.0",  
    "python-dotenv>=1.0.0",
    "uvicorn>=0.23.0",
    "fastapi>=0.104.0",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.389", 
    "ruff>=0.7.3",
    "black>=23.9.1",
    "isort>=5.12.0"
]

[project.scripts]
mcp-ssh-exec = "mcp_ssh_exec:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_ssh_exec"]
